resume_extractor/requirements.txt
confluent-kafka==2.3.0  
psycopg2-binary==2.9.9  
requests==2.32.3  
textract==1.6.5       
python-dotenv==1.0.1    
cloudinary
pdf2image
pypdf